divide&conquer
