+++ js/data/game-data.js
@@ -12,4 +12,28 @@ const ImgType = {
 };
 
 
-export {GameKind, ImgType};
+const AnswerType = {
+  FAST: `FAST`,
+  SLOW: `SLOW`,
+  WRONG: `WRONG`,
+  CORRECT: `CORRECT`,
+  UNKNOWN: `UNKNOWN`,
+};
+
+
+const TimeLine = {
+  [AnswerType.FAST]: 20,
+  [AnswerType.SLOW]: 10
+};
+
+
+const ScoreBonus = {
+  CORRECT: 100,
+  LIVES: 50,
+  [AnswerType.FAST]: 50,
+  [AnswerType.SLOW]: -50,
+  [AnswerType.WRONG]: 0,
+  [AnswerType.UNKNOWN]: 0
+};
+
+export {GameKind, ImgType, TimeLine, ScoreBonus, AnswerType};
