+++ js/data/game-mechanics.js
@@ -1,5 +1,6 @@
-import {GameSetting, AnswerType, TimeLine} from './../game-rules.js';
-import {GameKind} from './game-data.js';
+import {GameSetting} from './game-setting.js';
+import {GameKind, AnswerType, TimeLine, ScoreBonus} from './game-data.js';
+
 
 const INITIAL_GAME = Object.freeze({
   level: 0,
@@ -9,6 +10,33 @@ const INITIAL_GAME = Object.freeze({
   time: GameSetting.TIME_LIMIT,
   game: null
 });
+
+
+// @param {answers} array on answer objects
+// @param {livesAmount} amount of lives at the end of the game
+// $return scores if win or -1 if lost
+
+const getScores = (answers, livesAmount) => {
+  if (!Array.isArray(answers) || answers.some((it) => it !== Object(it))) {
+    throw new Error(`the answers parameter should be array of objects`);
+  }
+
+  if (typeof livesAmount !== `number` || isNaN(livesAmount)) {
+    throw new Error(`the lives parameter should be a number`);
+  }
+
+  if (livesAmount > GameSetting.INITIAL_LIVES) {
+    throw new Error(`the lives can't be higher than ${GameSetting.INITIAL_LIVES}`);
+  }
+
+  if (answers.length !== GameSetting.MAX_LEVEL || livesAmount < 0) {
+    return -1;
+  }
+
+  return answers.reduce((accumulator, current) => {
+    return accumulator + ScoreBonus[current.type] + (current.isCorrect && current.type !== AnswerType.CORRECT ? ScoreBonus.CORRECT : 0);
+  }, livesAmount * ScoreBonus.LIVES);
+};
 
 
 const createUserAnswer = (answerStatus, time) => {
@@ -84,4 +112,4 @@ const reapLife = (state) => {
 };
 
 
-export {INITIAL_GAME, changeLevel, reapLife, createUserAnswer, checkUserAnswer, getUserAnswers};
+export {INITIAL_GAME, changeLevel, reapLife, createUserAnswer, checkUserAnswer, getUserAnswers, getScores};
