+++ js/screen-game/screen-game-model.js
@@ -1,5 +1,6 @@
-import {INITIAL_GAME, changeLevel, reapLife} from './data/game-mechanics.js';
-import {tick, resetTimer} from './timer.js';
+import {INITIAL_GAME, changeLevel, reapLife} from './../data/game-mechanics.js';
+import {tick, resetTimer} from './../utils/tick.js';
+
 
 export default class GameModel {
   constructor(data) {
