+++ js/screen-rules/screen-rules-view.js
@@ -1,4 +1,4 @@
-import AbstractView from './abstract-view.js';
+import AbstractView from './../abstract-view/abstract-view.js';
 
 
 const InteractiveElementSelector = {
