+++ js/screen-rules/screen-rules.js
@@ -1,4 +1,4 @@
-import Application from './application.js';
+import Application from './../application.js';
 import ScreenRulesView from './screen-rules-view.js';
 
 export default class ScreenRules {
