+++ js/screen-stats/screen-stats.js
@@ -3,7 +3,6 @@ import ScreenStatsView from './screen-stats-view.js';
 export default class ScreenStats {
   constructor(state) {
     this.root = new ScreenStatsView(state);
-    // this.root.addResults();
   }
 
   get element() {
