+++ js/test/utils.test.js
@@ -1,5 +1,37 @@
 import assert from 'assert';
-import resizeImg from './../resize-img.js';
+import resizeImg from './../utils/resize-img.js';
+import {tick} from './../utils/tick.js';
+
+
+describe(`Testing updateTime() - function to change timer in game state object`, () => {
+  // incorrect data
+
+  it(`should deal with incorrect data inside state object`, () => {
+    assert.throws(() => tick({time: -1}), /incorrect data, state object's time property should be in interval from 0 to 30/);
+    assert.throws(() => tick({time: 31}), /incorrect data, state object's time property should be in interval from 0 to 30/);
+  });
+
+  // corner cases
+
+  it(`should return object with updated time property`, () => {
+    assert.equal(tick({time: 30}).time, 29);
+    assert.equal(tick({time: 1}).time, 0);
+  });
+
+  it(`should not increase level if state object time is 0`, () => {
+    assert.equal(tick({time: 0}).time, 0);
+  });
+
+  // invalid data
+
+  it(`it should only take object with level property as a parameter`, () => {
+    assert.throws(() => tick({}), /is not an object or has no time property/);
+    assert.throws(() => tick([]), /is not an object or has no time property/);
+    assert.throws(() => tick(0), /is not an object or has no time property/);
+    assert.throws(() => tick({notValidProperty: 2}), /is not an object or has no time property/);
+  });
+});
+
 
 const createTestForFrame = (frame) => {
   const assertRatio = (given, expected) => {
